fix: handle not expected errors in testCases

+++ e2e/toDeleteRole/test2.spec.ts
@@ -2,5 +2,8 @@
 
 it("should fail in delete a role", async () => {
   const role = corde.bot.getRole({ name: "role-to-delete" });
+  if (!role) {
+    return corde.fail();
+  }
   await command("deleteRole 1231").should.deleteRole(role.id);
 });
